feat(ISV-5875): handle retry exhaustion error separately
We need to return a different exit code for transient Atlas errors and
actual errors in the script. This is because we don't want to fail the
release Tekton task when a transient error occurs (we have the S3
retry), but we want it to fail when an unexpected error occurs.

Signed-off-by: Martin Jediny <jedinym@proton.me>