feat: Simplify the Github release process
The release-please is capable to drive the Github release with
additional command. The command replaces the manual steps that we had to
do directly with git or Github.

The provided Makefile simplifies the process by running just 2 commands.
One that opens a release PR and other that creates a release.

Signed-off-by: Ales Raszka <araszka@redhat.com>