feat: reduce image arch for pull request to x86_64
I realized that running a build for each pull request makes the pipeline
unnecessary long and we don't need the extra arches to be build for
every PR.

This commit removes extra arches for pull request but keeps all arches
for push pipelines. This make sure we still produce multi arch images
for a Mobster releases.

Signed-off-by: Ales Raszka <araszka@redhat.com>