test(ISV-6206): A new integration test to vefify SBOM with Conforma
A new test aims to verify the SBOMs produced by the Mobster is aligned
with the Conforma rules.

A test generates image, index image and generates SBOM for both.
In order to verify the Conforma rules both images needs to be signed and
attestation needs to be provided. The attestation is dynamically
generate based on what is being produced by Konflux build.

JIRA: ISV-6206

The test discovered an issue in Conforma that is addressed here:
https://github.com/conforma/policy/pull/1531

Signed-off-by: Ales Raszka <araszka@redhat.com>