fix: replace erroneous uses of "enrich" with "augment"
Enrichment and augmentation are two distinct processes that can be
performed on an SBOM.

Today, mobster only knows how to augment. It does not know how to
enrich.

In the future, we want to add enrichment as a separate process.

This change revises erroneous uses of "enrichment" in the codebase to
more accurately say "augmentation" to avoid future confusion.